Refine architecture
The architecture of the program is now refined, with platorm specific code fully in a seperate file, and different jobs clearly deliminated. Filenames have been updated to convey the difference. chip8_core.py defines the State class, and a new Interpreter class which executes opcodes, found in interpreter_opcodes.py, on instances of State. chip8.py is now the main executable. It uses a platfrom layer, which defines the class Interface. The interface contains references to Audio, Video and Input classes. Currently the only platform is SDL, which should be compatible with all OSes. It should be simple to add new platfroms.