Ease debugging, and fix bugs. Fulfill top two tasks in TODO.
Minor changes to a lot of functions to ease debugging. The structure is now very much about handing a State object round which represents the current state of the virtual machine. chip8.chip8() (the main loop) now optionally takes a State object. This all eases debugging, as minimal amounts have to be entered into an interactive Python session to get started. The change to chip8() means it's possible to run it with a State object from the interactive shell, access and change it whenever you exit the main loop, and then pass it back in again and continue where you left off.

The easier debugging helped with the discovery of two bugs. One was with opcodes, a difference between how the original interpreter did it, and how modern interpreters (and so relatively modern code) do it. The other was with the delay timer.